+++ internal/db/models.go
@@ -65,6 +65,7 @@ type Server struct {
 	PortMax         int    `json:"port_max"`
 	ExpiresAt       int64  `json:"expires_at"`
 	TrafficResetDay int    `json:"traffic_reset_day"`
+	DisableIPv6     bool   `json:"disable_ipv6"`
 	CanPushCores    bool   `json:"can_push_cores,omitempty"`
 }